moved db schema check for more general package and added it into pt also + minor dev log conf tuning
