+++ exposure-notification/src/main/java/fi/thl/covid19/exposurenotification/efgs/dao/OutboundOperationDao.java
@@ -3,7 +3,6 @@ package fi.thl.covid19.exposurenotification.efgs.dao;
 import fi.thl.covid19.exposurenotification.efgs.entity.OutboundOperation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
@@ -117,16 +116,5 @@ public class OutboundOperationDao {
                         "datetime", Timestamp.valueOf(LocalDateTime.now(ZoneOffset.UTC).minus(1, DAYS))
                 ),
                 (rs, i) -> rs.getInt(1)).stream().findFirst().orElseThrow(() -> new IllegalStateException("Count returned nothing."));
-    }
-
-    @Transactional
-    public boolean checkEnSchemaExists() {
-        String sql = "select count(*) from information_schema.schemata where schema_name = :schema_name";
-        try {
-            return jdbcTemplate.query(sql, Map.of("schema_name", "en"),
-                    (rs, i) -> rs.getInt(1)).stream().findFirst().orElseThrow(() -> new IllegalStateException("Count returned nothing.")) > 0;
-        } catch (DataAccessException | IllegalStateException e) {
-            return false;
-        }
     }
 }
